refactor: auth, member 분리 및 비밀번호 검증을 member의 책임으로 변경
