+++ src/main/java/roomescape/application/AuthService.java
@@ -1,0 +1,36 @@
+package roomescape.application;
+
+import org.springframework.stereotype.Service;
+import roomescape.application.dto.request.TokenCreationRequest;
+import roomescape.application.dto.response.MemberResponse;
+import roomescape.application.dto.response.TokenResponse;
+import roomescape.auth.TokenProvider;
+import roomescape.domain.member.Member;
+import roomescape.domain.member.MemberRepository;
+
+@Service
+public class AuthService {
+    private static final String WRONG_EMAIL_OR_PASSWORD_MESSAGE = "등록되지 않은 이메일이거나 비밀번호가 틀렸습니다.";
+
+    private final TokenProvider tokenProvider;
+    private final MemberRepository memberRepository;
+
+    public AuthService(TokenProvider tokenProvider, MemberRepository memberRepository) {
+        this.tokenProvider = tokenProvider;
+        this.memberRepository = memberRepository;
+    }
+
+    public TokenResponse authenticateMember(TokenCreationRequest request) {
+        Member member = memberRepository.findByEmail(request.email())
+                .orElseThrow(() -> new IllegalArgumentException(WRONG_EMAIL_OR_PASSWORD_MESSAGE));
+        member.validatePassword(request.password());
+        String token = tokenProvider.createToken(Long.toString(member.getId()));
+        return new TokenResponse(token);
+    }
+
+    public MemberResponse getMemberById(long id) {
+        Member member = memberRepository.findById(id)
+                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 회원입니다."));
+        return MemberResponse.from(member);
+    }
+}
