+++ src/test/java/roomescape/support/ControllerTest.java
@@ -5,7 +5,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.web.servlet.MockMvc;
-import roomescape.application.MemberService;
+import roomescape.application.AuthService;
 import roomescape.application.ReservationService;
 import roomescape.application.ReservationTimeService;
 import roomescape.application.ThemeService;
@@ -34,7 +34,7 @@ public class ControllerTest {
     @MockBean
     private ThemeService themeService;
     @MockBean
-    private MemberService memberService;
+    private AuthService authService;
     @MockBean
     private AuthenticationInfoExtractor authenticationInfoExtractor;
     @MockBean
