+++ src/test/java/roomescape/ui/controller/AuthControllerTest.java
@@ -9,7 +9,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.web.servlet.ResultActions;
-import roomescape.application.MemberService;
+import roomescape.application.AuthService;
 import roomescape.application.dto.response.TokenResponse;
 import roomescape.support.ControllerTest;
 import roomescape.support.SimpleMockMvc;
@@ -17,11 +17,11 @@ import roomescape.ui.controller.dto.LoginRequest;
 
 class AuthControllerTest extends ControllerTest {
     @Autowired
-    private MemberService memberService;
+    private AuthService authService;
 
     @Test
     void 로그인을_성공한다() throws Exception {
-        when(memberService.authenticateMember(any())).thenReturn(new TokenResponse("token!!!"));
+        when(authService.authenticateMember(any())).thenReturn(new TokenResponse("token!!!"));
         LoginRequest request = new LoginRequest("abc@gmail.com", "1q2w3e4r!");
         String content = objectMapper.writeValueAsString(request);
